Prevent layout from breaking on super low width
The static minimum values for font sizes caused the text contents to overflow
their containers in very low width screens. Preventing that required the font
sizes to be relative to the screen width, without a minimum size. This prevents
the layout from breaking, but at the same time causes the fonts to be extremely
small, unreadable in low width views.